feat: improve the flexibility of T type, KdTree.size. Add generate nearest_one_point
o Relax the trait constraints on KdTree content type T and use T::default()
o Make KdTree.size a usize instead of a T, simplifying interface and semantics
o To allow use of the K-D Tree coordinates, provide nearest_one_point (in
  addition to nearest_one), which returns the NearestNeighbour and also the
  coordinate of the nearest K-D Tree point.

+++ src/float/query/best_n_within.rs
@@ -196,7 +196,7 @@ mod tests {
         content_to_add
             .iter()
             .for_each(|(point, content)| tree.add(point, *content));
-        assert_eq!(tree.size(), TREE_SIZE as i32);
+        assert_eq!(tree.size(), TREE_SIZE);
 
         let query_points: Vec<[AX; 2]> = (0..NUM_QUERIES)
             .map(|_| rand::random::<[AX; 2]>())
